// src/types/contents/education.ts
export type EducationEntry = {
  id: string // '76de100f-3eae-4ef2-9bd7-f1b51236df65';
  isHidden: boolean // false;
  showPlaceholder: boolean // false;
  createdAt: string // '2026-07-01T14:26:22.024Z';
  updatedAt: string // '2026-07-01T14:28:23.707Z';

  degree?: string // 'graduate';
  school?: string // 'a school name';
  location?: string // 'ny';
  endDateNew?: string // 'Present';
  schoolLink?: string // '';
  description?: string // html: '<p>this is my school</p>';
  startDateNew?: string // '01/2020';
}